Update snakemake for DGN
add postanalysis to pipeline
modified z.sh

+++ script/z.sh
@@ -5,6 +5,8 @@ prefix=$4
 dir_script=$5
 out=$6
 
+set -o nounset -o errexit -o pipefail
+
 module load R/3.6.1
 #source activate /home/liliw1/.conda/envs/venv
 
@@ -27,17 +29,23 @@ if (( nBatch >= 1 )); then
     --maf_threshold 0 \
     --output_text
     
-    zcat $prefix_tmp'.trans_qtl_pairs.txt.gz' | \
-    tail -n+2 | \
-    awk '{printf "%s\t%s\t%.6f\n", $1,$2,$4/$5}' \
-    > $prefix_tmp'.trans_qtl_pairs_z.txt'
-    rm $prefix_tmp'.trans_qtl_pairs.txt.gz'
-    rm $prefix_tmp'.tensorQTL.trans.log'
+    if [[ -f $prefix_tmp'.trans_qtl_pairs.txt.gz' ]]
+    then
+      zcat $prefix_tmp'.trans_qtl_pairs.txt.gz' | \
+      tail -n+2 | \
+      awk '{printf "%s\t%s\t%.6f\n", $1,$2,$4/$5}' \
+      > $prefix_tmp'.trans_qtl_pairs_z.txt'
+      rm $prefix_tmp'.trans_qtl_pairs.txt.gz'
+      rm $prefix_tmp'.tensorQTL.trans.log'
+    fi
   done
   
-  cat $prefix'.'*'.trans_qtl_pairs_z.txt' | awk 'BEGIN{print "snp\tgene\tzscore"}{print $0}' > \
-  $prefix'.trans_qtl_pairs_z.txt'
-  rm $prefix'.'*'.trans_qtl_pairs_z.txt'
+  nsucBatch=$(ls $prefix.[1-$nBatch].trans_qtl_pairs_z.txt | wc -l)
+  if ((nsucBatch == nBatch)); then
+    cat $prefix.[1-$nBatch].trans_qtl_pairs_z.txt | awk 'BEGIN{print "snp\tgene\tzscore"}{print $0}' > \
+    $prefix.trans_qtl_pairs_z.txt
+    rm $prefix.[1-$nBatch].trans_qtl_pairs_z.txt
+  fi
   
 else
   ~/.conda/envs/venv/bin/python3 -m tensorqtl \
@@ -48,17 +56,20 @@ else
   --maf_threshold 0 \
   --output_text
   
-  zcat $prefix'.trans_qtl_pairs.txt.gz' | \
-  tail -n+2 | \
-  awk 'BEGIN{print "snp\tgene\tzscore"}{printf "%s\t%s\t%.6f\n", $1,$2,$4/$5}' \
-  > $prefix'.trans_qtl_pairs_z.txt'
-  rm $prefix'.trans_qtl_pairs.txt.gz'
-  rm $prefix'.tensorQTL.trans.log'
+  if [[ -f $prefix'.trans_qtl_pairs.txt.gz' ]]
+  then
+    zcat $prefix'.trans_qtl_pairs.txt.gz' | \
+    tail -n+2 | \
+    awk 'BEGIN{print "snp\tgene\tzscore"}{printf "%s\t%s\t%.6f\n", $1,$2,$4/$5}' \
+    > $prefix'.trans_qtl_pairs_z.txt'
+    rm $prefix'.trans_qtl_pairs.txt.gz'
+    rm $prefix'.tensorQTL.trans.log'
+  fi
 fi
 
 # convert z.txt to z matrix
 Rscript --no-save --no-restore $dir_script'make.zmat.R' $prefix".trans_qtl_pairs_z.txt" $out
-rm $prefix'.trans_qtl_pairs_z.txt'
+rm $prefix.trans_qtl_pairs_z.txt
 
 #conda deactivate
